testing: Added Code Coverage Analysis EclEmma
