+++ .classpath
@@ -1,10 +1,25 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry exported="true" kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER"/>
+	<classpathentry kind="src" output="target/classes" path="src">
+		<attributes>
+			<attribute name="optional" value="true"/>
+			<attribute name="maven.pomderived" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8">
+		<attributes>
+			<attribute name="maven.pomderived" value="true"/>
+		</attributes>
+	</classpathentry>
 	<classpathentry kind="lib" path="C:/Users/eunbee/Desktop/college/S5/OSD/Release/findBrokenGlass/lib/jna-5.6.0.jar"/>
 	<classpathentry kind="lib" path="C:/Users/eunbee/Desktop/college/S5/OSD/Release/findBrokenGlass/lib/json-simple-1.1.jar"/>
 	<classpathentry kind="lib" path="C:/Users/eunbee/Desktop/college/S5/OSD/Release/findBrokenGlass/lib/jsoup-1.13.1.jar"/>
 	<classpathentry kind="lib" path="C:/Users/eunbee/Desktop/college/S5/OSD/Release/findBrokenGlass/lib/google-java-format-1.9-all-deps.jar"/>
-	<classpathentry kind="output" path="bin"/>
+	<classpathentry kind="con" path="org.eclipse.jdt.junit.JUNIT_CONTAINER/5"/>
+	<classpathentry kind="con" path="org.eclipse.m2e.MAVEN2_CLASSPATH_CONTAINER">
+		<attributes>
+			<attribute name="maven.pomderived" value="true"/>
+		</attributes>
+	</classpathentry>
+	<classpathentry kind="output" path="target/classes"/>
 </classpath>
