+++ formatRun.bat
@@ -1,4 +1,5 @@
 @echo Formatter Script start....
-@java -jar lib/google-java-format-1.9-all-deps.jar --replace src/*.java
+@java -jar lib/google-java-format-1.9-all-deps.jar --replace src/main/*.java
+@java -jar lib/google-java-format-1.9-all-deps.jar --replace src/test/*.java
 @echo Done
 @pause
