+++ index.html
@@ -1,68 +1,151 @@
-#Feed excluded due to unable to verify the first certificate
-#[http://hcoelho.com/blog/feed]
-#name=Henrique Coelho
+
+[http://hendrikinmozilla.wordpress.com/feed/]
+name=Hendrik
 
 
-#Feed excluded due to Request failed with status code 404
-#[https://jalemansite.wordpress.com/category/feed/]
-#name=Johann Aleman
+[http://igoryen.wordpress.com/category/CDOT/feed/]
+name=Igor Entaltsev
 
 
-#Feed excluded due to Request failed with status code 410
-#[http://sinomai.wordpress.com/category/planet-cdot/feed/]
-#name=Matt Jang
+[https://mrrajevski.wordpress.com/category/spo600/feed]
+name=Matt Rajevski
 
 
-#Feed excluded due to Request failed with status code 410
-#[http://hzahoori.wordpress.com/category/open-source/feed]
-#name=Habib Zahoori
+[http://yuriykartuzov.wordpress.com/category/Open%20Source/feed/]
+name=Yuriy Kartuzov
 
 
-#Feed excluded due to Request failed with status code 404
-#[http://vprasanth.com/blog/archives/category/spo600-2/feed]
-#name=Prasanth Vaaheeswaran
+[https://medium.com/feed/@alexei.kozachenko]
+name=Aleksei Kozachenko
 
 
-#Feed excluded due to Request failed with status code 410
-#[http://rlawrence5.wordpress.com/feed/]
-#name=Ryan Lawrence
+[http://minghans.wordpress.com/category/spo600/feed/]
+name=Minghan Xu
 
 
-#Feed excluded due to Request failed with status code 404
-#[https://apspo600.blogspot.com/feed/posts/default/]
-#name=Atilla Puskas
+[http://mbustossbr.wordpress.com/feed/]
+name=Maria Bustos-Roman
 
 
-#Feed excluded due to Request failed with status code 410
-#[https://bapetov.wordpress.com/category/opensourcedps909/feed]
-#name=Bakytzhan Apetov
+[http://alicoding.com/rss/]
+name=Ali Al Dallal
 
 
-#Feed excluded due to Request failed with status code 410
-#[http://adsantokhi.wordpress.com/category/open-source/feed]
-#name=Anil Santokhi
+[http://okhattab.wordpress.com/feed/]
+name=Omarr Khattab
 
 
-#Feed excluded due to Request failed with status code 404
-#[http://rawkamatic.github.io/opensourcefeed]
-#name=Hunter Jansen
+[http://jsinghfoss.wordpress.com/category/spring-framework/feed/]
+name=Jatinder Singh
 
 
-#Feed excluded due to Request failed with status code 404
-#[http://ayufidin.blogspot.ca/feeds/posts/default]
-#name=Alon Yufidin
+[http://franksun123.blogspot.com/feeds/posts/default/]
+name=Dong Sun
 
 
-#Feed excluded due to Request failed with status code 404
-#[http://kevinpaiva.com/category/spo600/feed/]
-#name=Kevin Manuel Paiva
+[http://maximumou.blogspot.com/feeds/posts/default]
+name=Zhi Chang Ou
 
 
-#Feed excluded due to Request failed with status code 410
-#[http://dboddie46.wordpress.com/category/SBR600A/feed/]
-#name=Derrick Boddie
+[http://scottosd.blogspot.com/feeds/posts/default/-/Open%20Source/]
+name=Scott Lunel
 
 
-#Feed excluded due to Request failed with status code 404
-#[http://liandrew.ca/feed.opensource.xml]
-#name=Andrew Li
+[https://mb30myopensourceblog.blogspot.ca/feeds/posts/default]
+name=Patrick Godbout
+
+
+[https://opensourcetoronto.wordpress.com/category/open-source/feed/]
+name=Hans van den Pol
+
+
+[https://dcchen.home.blog/category/spo600/feed/]
+name=Danny Chen
+
+
+[http://gchau2.wordpress.com/feed/]
+name=Gary Chau
+
+
+[https://timothymoy.wordpress.com/tag/open-source/feed/]
+name=Timothy Moy
+
+
+[http://capereir.wordpress.com/feed/]
+name=Chris Pereira
+
+
+[http://dbaranski.wordpress.com/category/open-source/feed/]
+name=Dominic Baranski
+
+
+[http://enderstruth.wordpress.com/category/open-source/feed/]
+name=Roger Dicke
+
+
+[http://senecahealth.wordpress.com/feed/]
+name=Seneca Health Projects Blog
+
+
+[http://petepabs.wordpress.com/feed/]
+name=Peter Valerio
+
+
+[http://andrewow.wordpress.com/category/sbr600/feed/]
+name=Andrew Oatley-Willis
+
+
+[http://vlam6.wordpress.com/category/open-source/feed/]
+name=Vincent Lam
+
+
+[http://scorchedicee.wordpress.com/category/Seneca/feed/]
+name=Adam Sone
+
+
+[http://thevakaran.wordpress.com/category/open-source/feed/]
+name=Thevakaran Virutthasalam
+
+
+[http://dee132.blogspot.com/feeds/posts/default/-/seneca]
+name=Chris Bishop
+
+
+[https://brettlarney.wordpress.com/category/open-source/feed/]
+name=Brett Larney
+
+
+[http://dylansegna.wordpress.com/feed/]
+name=Dylan Segna
+
+
+[http://ben1amin.wordpress.com/category/seneca/feed/]
+name=Benjamin Chalovich
+
+
+[http://elnushaj.wordpress.com/feed/]
+name=Elsi Nushaj
+
+
+[https://ekatdev.wordpress.com/category/SPO600/feed/]
+name=Ekaterina Zaytseva
+
+
+[http://kopay.wordpress.com/category/sbr600-win2011/feed]
+name=Pirathapan Sivalingam
+
+
+[http://jessefulton.wordpress.com/category/SBR600/feed/]
+name=Jesse Fulton
+
+
+[http://eric-spo600.blogspot.com/feeds/posts/default]
+name=Eric Ferguson
+
+
+[http://armenzg.blogspot.com/feeds/posts/default/-/open%20source]
+name=Armen Zambrano G. (armenzg)
+
+
+[https://vishnuhacks.wordpress.com/category/spo600/feed/]
+name=Vishnu Santhosh Kumar
