+++ src/test/T001_VersionCheck.java
@@ -1,16 +1,16 @@
 package test;
 
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertTrue;
 import java.io.IOException;
 import org.junit.jupiter.api.Test;
+import junit.framework.TestCase;
 import main.UrlCheck;
 
-class T001_VersionCheck {
+class T001_VersionCheck extends TestCase {
 
   @Test
   @SuppressWarnings("static-access")
   void testMain() {
+    System.out.println("Test01");
     String[] args = {"-v"};
 
     UrlCheck urlCheck = new UrlCheck();
@@ -24,5 +24,4 @@ class T001_VersionCheck {
     }
     assertTrue(true);
   }
-
 }
