+++ src/test/T002_ArchievedCheck.java
@@ -24,5 +24,4 @@ class T002_ArchievedCheck {
     }
     assertTrue(true);
   }
-
 }
