+++ src/test/T003_HttpsRequestCheck.java
@@ -2,9 +2,10 @@ package test;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
+
 import java.io.IOException;
-import org.junit.jupiter.api.Test;
 import main.UrlCheck;
+import org.junit.jupiter.api.Test;
 
 class T003_HttpsRequestCheck {
 
@@ -25,5 +26,4 @@ class T003_HttpsRequestCheck {
 
     assertTrue(true);
   }
-
 }
