+++ src/test/T006_GetUrlFromTelescopeCheck.java
@@ -2,8 +2,9 @@ package test;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
-import org.junit.jupiter.api.Test;
+
 import main.UrlCheck;
+import org.junit.jupiter.api.Test;
 
 class T006_GetUrlFromTelescopeCheck {
 
@@ -32,7 +33,5 @@ class T006_GetUrlFromTelescopeCheck {
     } catch (RuntimeException e) {
       assertFalse(false);
     }
-
   }
-
 }
