+++ src/test/T007_visitFileRecursiveCheck.java
@@ -2,10 +2,11 @@ package test;
 
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
+
 import java.io.IOException;
 import java.util.ArrayList;
-import org.junit.jupiter.api.Test;
 import main.UrlCheck;
+import org.junit.jupiter.api.Test;
 
 class T007_visitFileRecursiveCheck {
 
@@ -48,5 +49,4 @@ class T007_visitFileRecursiveCheck {
 
     assertTrue(actual.size() == expected);
   }
-
 }
